Refactor: 도메인 디렉토리로 파일 위치 변경
