+++ src/main/java/com/join/core/history/service/ViewHistoryReadService.java
@@ -2,7 +2,7 @@ package com.join.core.history.service;
 
 import com.join.core.avatar.domain.Avatar;
 import com.join.core.avatar.domain.AvatarReader;
-import com.join.core.bookmark.service.BookmarkReader;
+import com.join.core.bookmark.domain.BookmarkReader;
 import com.join.core.enrollment.service.EnrollmentReader;
 import com.join.core.history.dto.response.ViewStudyReadResponse;
 import com.join.core.history.mapper.ViewHistoryMapper;
